utils/string-functions - rename string-format; change truncate into an arrow function
